requirements.txt
python-dotenv~=1.0.0
scrapfly-sdk[scrapy]
pytesseract
Pillow
